Publish DEPS for Chromium 18.0.1025.169
Incrementing VERSION to 18.0.1025.169

git-svn-id: svn://svn.chromium.org/chrome/branches/1025_166/src@134785 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/base/default_origin_bound_cert_store.cc
@@ -1,9 +1,10 @@
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 #include "net/base/default_origin_bound_cert_store.h"
 
+#include "base/bind.h"
 #include "base/message_loop.h"
 
 namespace net {
@@ -16,17 +17,21 @@ DefaultOriginBoundCertStore::DefaultOriginBoundCertStore(
     : initialized_(false),
       store_(store) {}
 
-void DefaultOriginBoundCertStore::FlushStore(Task* completion_task) {
+void DefaultOriginBoundCertStore::FlushStore(
+    const base::Closure& completion_task) {
   base::AutoLock autolock(lock_);
 
   if (initialized_ && store_)
     store_->Flush(completion_task);
-  else if (completion_task)
+  else if (!completion_task.is_null())
     MessageLoop::current()->PostTask(FROM_HERE, completion_task);
 }
 
 bool DefaultOriginBoundCertStore::GetOriginBoundCert(
     const std::string& origin,
+    SSLClientCertType* type,
+    base::Time* creation_time,
+    base::Time* expiration_time,
     std::string* private_key_result,
     std::string* cert_result) {
   base::AutoLock autolock(lock_);
@@ -38,6 +43,9 @@ bool DefaultOriginBoundCertStore::GetOriginBoundCert(
     return false;
 
   OriginBoundCert* cert = it->second;
+  *type = cert->type();
+  *creation_time = cert->creation_time();
+  *expiration_time = cert->expiration_time();
   *private_key_result = cert->private_key();
   *cert_result = cert->cert();
 
@@ -46,14 +54,19 @@ bool DefaultOriginBoundCertStore::GetOriginBoundCert(
 
 void DefaultOriginBoundCertStore::SetOriginBoundCert(
     const std::string& origin,
+    SSLClientCertType type,
+    base::Time creation_time,
+    base::Time expiration_time,
     const std::string& private_key,
     const std::string& cert) {
   base::AutoLock autolock(lock_);
   InitIfNecessary();
 
   InternalDeleteOriginBoundCert(origin);
-  InternalInsertOriginBoundCert(origin,
-                                new OriginBoundCert(origin, private_key, cert));
+  InternalInsertOriginBoundCert(
+      origin,
+      new OriginBoundCert(
+          origin, type, creation_time, expiration_time, private_key, cert));
 }
 
 void DefaultOriginBoundCertStore::DeleteOriginBoundCert(
@@ -63,34 +76,37 @@ void DefaultOriginBoundCertStore::DeleteOriginBoundCert(
   InternalDeleteOriginBoundCert(origin);
 }
 
-void DefaultOriginBoundCertStore::DeleteAll() {
+void DefaultOriginBoundCertStore::DeleteAllCreatedBetween(
+    base::Time delete_begin,
+    base::Time delete_end) {
   base::AutoLock autolock(lock_);
   InitIfNecessary();
   for (OriginBoundCertMap::iterator it = origin_bound_certs_.begin();
-       it != origin_bound_certs_.end(); ++it) {
-    OriginBoundCert* cert = it->second;
-    if (store_)
-      store_->DeleteOriginBoundCert(*cert);
-    delete cert;
+       it != origin_bound_certs_.end();) {
+    OriginBoundCertMap::iterator cur = it;
+    ++it;
+    OriginBoundCert* cert = cur->second;
+    if ((delete_begin.is_null() || cert->creation_time() >= delete_begin) &&
+        (delete_end.is_null() || cert->creation_time() < delete_end)) {
+      if (store_)
+        store_->DeleteOriginBoundCert(*cert);
+      delete cert;
+      origin_bound_certs_.erase(cur);
+    }
   }
-  origin_bound_certs_.clear();
+}
+
+void DefaultOriginBoundCertStore::DeleteAll() {
+  DeleteAllCreatedBetween(base::Time(), base::Time());
 }
 
 void DefaultOriginBoundCertStore::GetAllOriginBoundCerts(
-    std::vector<OriginBoundCertInfo>* origin_bound_certs) {
+    std::vector<OriginBoundCert>* origin_bound_certs) {
   base::AutoLock autolock(lock_);
   InitIfNecessary();
   for (OriginBoundCertMap::iterator it = origin_bound_certs_.begin();
        it != origin_bound_certs_.end(); ++it) {
-    OriginBoundCertInfo cert_info = {
-      it->second->origin(),
-      it->second->private_key(),
-      it->second->cert()
-    };
-    // TODO(rkn): Make changes so we can simply write
-    // origin_bound_certs->push_back(*it->second). This is probably best done
-    // by unnesting the OriginBoundCert class.
-    origin_bound_certs->push_back(cert_info);
+    origin_bound_certs->push_back(*it->second);
   }
 }
 
@@ -157,16 +173,6 @@ void DefaultOriginBoundCertStore::InternalInsertOriginBoundCert(
     store_->AddOriginBoundCert(*cert);
   origin_bound_certs_[origin] = cert;
 }
-
-DefaultOriginBoundCertStore::OriginBoundCert::OriginBoundCert() {}
-
-DefaultOriginBoundCertStore::OriginBoundCert::OriginBoundCert(
-    const std::string& origin,
-    const std::string& private_key,
-    const std::string& cert)
-    : origin_(origin),
-      private_key_(private_key),
-      cert_(cert) {}
 
 DefaultOriginBoundCertStore::PersistentStore::PersistentStore() {}
